requirements.txt
fastapi
uvicorn
openai
pydantic
streamlit
Pillow